refactor items to avoid code duplication
