+++ agonyforge-mud-demo/src/main/java/com/agonyforge/mud/demo/model/impl/MudItem.java
@@ -2,32 +2,19 @@ package com.agonyforge.mud.demo.model.impl;
 
 import com.agonyforge.mud.demo.model.constant.WearSlot;
 import jakarta.persistence.*;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.util.*;
 
 @Entity
 @Table(name = "mud_item")
-public class MudItem extends Persistent {
-    private static final Logger LOGGER = LoggerFactory.getLogger(MudItem.class);
-
+public class MudItem extends AbstractMudItem {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "instance_id")
-    private Long instanceId = null;
+    private Long instanceId;
     private Long id;
     private Long roomId;
     private Long chId;
-
-    @Transient
-    private List<String> transientNameList = new ArrayList<>();
-    private String nameList;
-    private String shortDescription;
-    private String longDescription;
-
-    @Convert(converter = WearSlot.Converter.class)
-    private EnumSet<WearSlot> wearSlots = EnumSet.noneOf(WearSlot.class);
 
     private WearSlot worn;
 
@@ -71,48 +58,6 @@ public class MudItem extends Persistent {
         this.roomId = roomId;
     }
 
-    public List<String> getNameList() {
-        if (!transientNameList.isEmpty()) {
-            return new ArrayList<>(transientNameList);
-        }
-
-        return nameList == null ? List.of() : Arrays.stream(nameList.split(",")).toList();
-    }
-
-    public void setNameList(List<String> names) {
-        transientNameList = new ArrayList<>(names);
-        nameList = String.join(",", transientNameList);
-    }
-
-    @PostLoad
-    private void loadNameList() {
-        transientNameList = nameList == null ? List.of() : Arrays.stream(nameList.split(",")).toList();
-    }
-
-    public String getShortDescription() {
-        return shortDescription;
-    }
-
-    public void setShortDescription(String shortDescription) {
-        this.shortDescription = shortDescription;
-    }
-
-    public String getLongDescription() {
-        return longDescription;
-    }
-
-    public void setLongDescription(String longDescription) {
-        this.longDescription = longDescription;
-    }
-
-    public EnumSet<WearSlot> getWearSlots() {
-        return wearSlots;
-    }
-
-    public void setWearSlots(EnumSet<WearSlot> wearSlots) {
-        this.wearSlots = wearSlots;
-    }
-
     public WearSlot getWorn() {
         return worn;
     }
@@ -125,11 +70,11 @@ public class MudItem extends Persistent {
     public boolean equals(Object o) {
         if (this == o) return true;
         if (!(o instanceof MudItem mudItem)) return false;
-        return Objects.equals(getId(), mudItem.getId());
+        return Objects.equals(getInstanceId(), mudItem.getInstanceId());
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(getId());
+        return Objects.hash(getInstanceId());
     }
 }
